TORCS/rl-torcs/requirements.txt
absl-py==0.9.0
astor==0.8.1
backports.weakref==1.0.post1
enum34==1.1.10
funcsigs==1.0.2
futures==3.3.0
gast==0.3.3
grpcio==1.28.1
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.16.6
opencv-python==4.2.0.32
pkg-resources==0.0.0
protobuf==3.11.3
six==1.14.0
sysv-ipc==1.0.1
tensorboard==1.11.0
tensorflow==1.15.2
termcolor==1.1.0
Werkzeug==1.0.1
